Renamed files to be more consistent
