+++ main/include/bmp280.hpp
@@ -1,3 +1,5 @@
 #pragma once
 
+#define BMPTAG "BMP280"
+
 void taskBMP280(void *pvParameters);
