+++ main/include/telemetry.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#define LORATAG "LORA"
+
 // LoRa packet struct
 typedef struct lora_packet{
   uint8_t packetid;
